+++ fj-doc-sample/src/test/java/test/org/fugerit/java/doc/sample/format/TestPOI.java
@@ -12,6 +12,7 @@ import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.xssf.usermodel.XSSFCellStyle;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import org.junit.Assert;
 import org.junit.Test;
 
 import test.org.fugerit.java.doc.sample.facade.BasicFacadeTest;
@@ -29,6 +30,7 @@ public class TestPOI extends TestFormatBase {
 	
 	@Test
 	public void testCellForegroundColor() throws Exception  {
+		boolean ok = false;
 		String test = "cell_foreground_color";
 		Workbook workbook = new XSSFWorkbook();
 		Sheet sheet = workbook.createSheet( test );
@@ -47,6 +49,9 @@ public class TestPOI extends TestFormatBase {
 		
 		cell.setCellStyle( style );
 		close( workbook, test );
+		ok = true;
+		Assert.assertTrue(ok);
 	}
+	
 
 }